Use double. Fixes repayment calculator
Fixes console output

+++ LoanDecider/Loan.cs
@@ -3,12 +3,12 @@
     public class Loan
     {
         private double rate;
-        private decimal monthlyRepayment;
-        private decimal totalRepayment;
+        private double monthlyRepayment;
+        private double totalRepayment;
         private double principal;
         private string lenderName;
 
-        public Loan(string lenderName, double principal, double rate, decimal monthlyRepayment, decimal totalRepayment)
+        public Loan(string lenderName, double principal, double rate, double monthlyRepayment, double totalRepayment)
         {
             this.rate = rate;
             this.monthlyRepayment = monthlyRepayment;
@@ -18,8 +18,8 @@
         }
 
         public double Rate { get => rate; }
-        public decimal MonthlyRepayment { get => monthlyRepayment;  }
-        public decimal TotalRepayment { get => totalRepayment;  }
+        public double MonthlyRepayment { get => monthlyRepayment;  }
+        public double TotalRepayment { get => totalRepayment;  }
         public double Principal { get => principal; }
         public string Lender { get => lenderName; }
     }
